backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "pydantic>=2.12.5",
    "google-cloud-storage>=2.14.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "ade-python>=0.2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
]

[tool.setuptools]
packages = ["app"]
